Refactor petals.client.remote_model and petals.bloom.modeling_utils
